# docs/swagger.yaml
basePath: /api
definitions:
  internal_domains_user.ChangePasswordRequest:
    properties:
      confirm_password:
        minLength: 8
        type: string
      new_password:
        minLength: 8
        type: string
      old_password:
        minLength: 8
        type: string
    required:
    - confirm_password
    - new_password
    - old_password
    type: object
  internal_domains_user.LoginRequest:
    properties:
      password:
        minLength: 8
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
  internal_domains_user.RegisterRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 8
        type: string
      phone:
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - username
    type: object
  internal_domains_user.UpdateProfileRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      phone:
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    type: object
host: localhost:5000
info:
  contact: {}
  description: Zyosa Api with Go Fiber
  title: Zyosa API
  version: "1.0"
paths:
  /:
    get:
      consumes:
      - application/json
      description: Returns a simple hello message
      produces:
      - application/json
      responses:
        "200":
          description: Hello, World!
          schema:
            type: string
      summary: Root Endpoint
      tags:
      - Root
  /user/login:
    post:
      consumes:
      - application/json
      description: Authenticates user and returns an access token
      parameters:
      - description: User login data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_domains_user.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login success
          schema:
            type: string
        "401":
          description: Username or password is incorrect
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Login user
      tags:
      - Auth
  /user/logout:
    post:
      description: Logs out the user by clearing the access token
      responses:
        "200":
          description: Logout success
          schema:
            type: string
      summary: Logout user
      tags:
      - Auth
  /user/profile:
    get:
      description: Retrieves the profile of the authenticated user
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Updates the authenticated user's profile information
      parameters:
      - description: User profile update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_domains_user.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Username or email is already in use
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - User
  /user/profile/password:
    put:
      consumes:
      - application/json
      description: Changes the authenticated user's password
      parameters:
      - description: Password change request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_domains_user.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password changed successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: New password and confirm password must be the same
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Change user password
      tags:
      - User
  /user/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account
      parameters:
      - description: User registration data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_domains_user.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Username or email has been taken
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - Auth
swagger: "2.0"